Added various functions and main function
In main function, I created the process where the user will go through
in order to add an item to the cart. Also created toString function to
display the cart, AddItemVerification function to verify correct input
if user wants to add an item, PriceVerification function to check the
price of the item.